fix boarder and move file
